Get the key ID only during tainting
