+++ pkg/agent/client/client_test.go
@@ -882,7 +882,6 @@ func TestFetchJWTSVID(t *testing.T) {
 
 	issuedAt := time.Now().Unix()
 	expiresAt := time.Now().Add(time.Minute).Unix()
-	tok := "eyJhbGciOiJFUzI1NiIsImtpZCI6InlueVV3TEg5bklOaWcyWEdRdTBkdklWbndieG5JeHlPIiwidHlwIjoiSldUIn0.eyJhdWQiOlsidGVzdC1hdWRpZW5jZSJdLCJleHAiOjE3MjQzNjYwOTIsImlhdCI6MTcyNDI3OTc0Nywic3ViIjoic3BpZmZlOi8vZXhhbXBsZS5vcmcvYWdlbnQvZGJ1c2VyIn0.feVAZPLmWT4ohzNKBoil90WBn64noqnCYuXXJjhItdtsOmJqMBm-blfJl4pBvbRBCWym2YaXK9gl9RoAfFG0zQ"
 	for _, tt := range []struct {
 		name       string
 		setupTest  func(err error)
@@ -894,17 +893,16 @@ func TestFetchJWTSVID(t *testing.T) {
 			name: "success",
 			setupTest: func(err error) {
 				tc.svidServer.jwtSVID = &types.JWTSVID{
-					Token:     tok,
+					Token:     "token",
 					ExpiresAt: expiresAt,
 					IssuedAt:  issuedAt,
 				}
 				tc.svidServer.newJWTSVID = err
 			},
 			expectSVID: &JWTSVID{
-				Token:     tok,
+				Token:     "token",
 				ExpiresAt: time.Unix(expiresAt, 0).UTC(),
 				IssuedAt:  time.Unix(issuedAt, 0).UTC(),
-				Kid:       "ynyUwLH9nINig2XGQu0dvIVnwbxnIxyO",
 			},
 		},
 		{
@@ -927,7 +925,7 @@ func TestFetchJWTSVID(t *testing.T) {
 			name: "missing issuedAt",
 			setupTest: func(err error) {
 				tc.svidServer.jwtSVID = &types.JWTSVID{
-					Token:     tok,
+					Token:     "token",
 					ExpiresAt: expiresAt,
 				}
 				tc.svidServer.newJWTSVID = err
@@ -938,7 +936,7 @@ func TestFetchJWTSVID(t *testing.T) {
 			name: "missing expiredAt",
 			setupTest: func(err error) {
 				tc.svidServer.jwtSVID = &types.JWTSVID{
-					Token:    tok,
+					Token:    "token",
 					IssuedAt: issuedAt,
 				}
 				tc.svidServer.newJWTSVID = err
@@ -949,7 +947,7 @@ func TestFetchJWTSVID(t *testing.T) {
 			name: "issued after expired",
 			setupTest: func(err error) {
 				tc.svidServer.jwtSVID = &types.JWTSVID{
-					Token:     tok,
+					Token:     "token",
 					ExpiresAt: issuedAt,
 					IssuedAt:  expiresAt,
 				}
@@ -961,7 +959,7 @@ func TestFetchJWTSVID(t *testing.T) {
 			name: "grpc call to NewJWTSVID fails",
 			setupTest: func(err error) {
 				tc.svidServer.jwtSVID = &types.JWTSVID{
-					Token:     tok,
+					Token:     "token",
 					ExpiresAt: expiresAt,
 					IssuedAt:  issuedAt,
 				}
